feat: notify user of foreign key restricting deletion

+++ app/Http/Controllers/PositionController.php
@@ -147,6 +147,10 @@ class PositionController extends Controller
     public function destroy(Position $position)
     {
 
+        if ($position->incoming_figures()->count() > 0 || $position->outgoing_figures()->count() > 0) {
+            return back()->with('error', 'Deleting this position is intentionally restricted because one or more figures rely on this position. You could first delete all dependent figures, then delete this position.');
+        }
+
         DB::transaction(function () use ($position) {
             $position_family = $position->position_family;
             $position->delete();
